network update
now using bach normalization it is little bi faster
in the commented section after first set of convolutional layer we can add the side network(extra information), but im not sure if it is the best solution since we might have done something wrong while preparing side network. for now it is commented and we are using 64 feature maps. when we are using the side network we will use 16 feature maps just before adding the networks afterwards we can use more features